src/requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
urllib3==1.25.11
wrapt==1.12.1
python-telegram-bot 
python-decouple